* moved functions.py to classes dir
